pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "squorum"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.31",
  "pandas>=2.0",
  "numpy>=1.24",
  "python-dotenv>=1.0",
  "supabase>=2.0",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "squorum_core*"]
exclude = ["_deprecated*"]